+++ headeronly_src/sqlite3ppext.ipp
@@ -133,7 +133,7 @@ namespace sqlite3pp
 
     inline void context::result(char const* value, bool fcopy)
     {
-      sqlite3_result_text(ctx_, value, std::strlen(value), fcopy ? SQLITE_TRANSIENT : SQLITE_STATIC);
+      sqlite3_result_text(ctx_, value, static_cast<int>(std::strlen(value)), fcopy ? SQLITE_TRANSIENT : SQLITE_STATIC);
     }
 
     inline void context::result(void const* value, int n, bool fcopy)
@@ -158,7 +158,7 @@ namespace sqlite3pp
 
     inline void context::result_error(char const* msg)
     {
-      sqlite3_result_error(ctx_, msg, std::strlen(msg));
+      sqlite3_result_error(ctx_, msg, static_cast<int>(std::strlen(msg)));
     }
 
     inline void* context::aggregate_data(int size)
